updated refactoring
